+++ apps/web/src/modules/CourseSearch/components/FilterSection/utils/addTag/index.test.tsx
@@ -1,3 +1,5 @@
+import { describe, expect, it } from 'vitest'
+
 import { addTag } from '.'
 
 describe('addTag', () => {