+++ apps/admin-web/CHANGELOG.md
@@ -1,5 +1,29 @@
 # admin-web
 
+## 1.2.0-beta.1
+
+### Minor Changes
+
+- c0e5777: feat: next 14 and minor dependencies
+
+### Patch Changes
+
+- Updated dependencies [c0e5777]
+  - @cgr/codegen@1.4.0-beta.1
+  - @cgr/course-utils@1.2.0-beta.1
+
+## 1.2.0-beta.0
+
+### Minor Changes
+
+- 4f342ee: deps: upgrade to node 20 and bump all patch-minor
+
+### Patch Changes
+
+- Updated dependencies [4f342ee]
+  - @cgr/course-utils@1.2.0-beta.0
+  - @cgr/codegen@1.4.0-beta.0
+
 ## 1.1.0
 
 ### Minor Changes