+++ packages/codegen/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @cgr/codegen
 
+## 1.4.0-beta.1
+
+### Minor Changes
+
+- c0e5777: feat: next 14 and minor dependencies
+
+## 1.4.0-beta.0
+
+### Minor Changes
+
+- 4f342ee: deps: upgrade to node 20 and bump all patch-minor
+
 ## 1.3.0
 
 ### Minor Changes