+++ packages/project-config/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @cgr/project-config
 
+## 1.2.0-beta.1
+
+### Minor Changes
+
+- c0e5777: feat: next 14 and minor dependencies
+
+## 1.2.0-beta.0
+
+### Minor Changes
+
+- 4f342ee: deps: upgrade to node 20 and bump all patch-minor
+
 ## 1.1.0
 
 ### Minor Changes